[WFGP-224] Separating code needed to run forked WfConfigGenerator
